requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
chromadb==1.3.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.14
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
emoji==2.15.0
fastapi==0.121.1
filelock==3.20.0
filetype==1.2.0
flasgger==0.9.7.1
Flask==3.1.2
flask-sock==0.7.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.2.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-ollama==0.3.10
langchain-openai==0.3.35
langchain-redis==0.2.4
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.42
logistro==2.0.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
minio==7.2.18
mistune==3.1.4
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
narwhals==2.11.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.4
oauthlib==3.3.1
olefile==0.47
ollama==0.6.0
onnxruntime==1.23.2
openai==2.7.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
plotly==6.4.0
pluggy==1.6.0
ply==3.11
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pymilvus==2.6.3
PyMySQL==1.1.2
pypdf==6.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.11
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-ulid==3.1.0
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
redis==6.4.0
redisvl==0.11.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.18
unstructured-client==0.42.3
urllib3>=2.4.0,<3.0.0
uvicorn==0.38.0
vanna==0.7.9
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
wsproto==1.3.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
